Moved a file
TestDataGenerator is now in TestHelpers ns
